--- sdk/src/main/java/cloud/mindbox/mobile_sdk/pushes/MindboxMessageHandler.kt
@@ -1,6 +1,0 @@
-package cloud.mindbox.mobile_sdk.pushes
-
-internal data class MindboxMessageHandler(
-    val imageFailureHandler: MindboxImageFailureHandler = MindboxImageFailureHandler.Default,
-    val imageLoader: MindboxImageLoader = MindboxImageLoader.Default,
-)
