+++ sdk/src/main/java/cloud/mindbox/mobile_sdk/pushes/handler/image/ImageFailureHandlerDefaultStrategy.kt
@@ -1,0 +1,8 @@
+package cloud.mindbox.mobile_sdk.pushes.handler.image
+
+/**
+ * Default implementation of error handling on image loading
+ */
+fun MindboxImageFailureHandler.Companion.default(): MindboxImageFailureHandler = DefaultStrategy
+
+private val DefaultStrategy: MindboxImageFailureHandler = ApplyDefaultStrategyImpl()
