+++ sdk/src/main/java/cloud/mindbox/mobile_sdk/pushes/handler/image/ImageRetryStrategy.kt
@@ -1,4 +1,4 @@
-package cloud.mindbox.mobile_sdk.pushes
+package cloud.mindbox.mobile_sdk.pushes.handler.image
 
 import android.graphics.Bitmap
 import android.graphics.BitmapFactory
